feat: move code to src directory and update build and test scripts
This commit moves the code from the root directory to the src directory to better organize the project. It also updates the build and test scripts to reflect this change.

The package.json file has been updated with the following changes:
- Added "type": "module" to enable ES modules
- Updated "main" to point to the output directory
- Updated "files" to include only necessary files
- Updated "scripts" to reflect the new file locations
- Added "ava" configuration to enable ESM loading for tests

The tsconfig.json file has also been updated to include the new file locations.

The bin directory has been removed since the binary file is now created in the output directory.

This commit does not contain any breaking changes.